requirements.txt
beautifulsoup4==4.12.3
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
duckduckgo_search==7.3.0
filelock==3.17.0
frozendict==2.4.6
fsspec==2024.12.0
html5lib==1.1
huggingface-hub==0.28.0
idna==3.10
Jinja2==3.1.5
lxml==5.3.0
markdown-it-py==3.0.0
markdownify==0.14.1
MarkupSafe==3.0.2
mdurl==0.1.2
multitasking==0.0.11
numpy==2.2.2
packaging==24.2
pandas==2.2.3
peewee==3.17.8
pillow==11.1.0
platformdirs==4.3.6
primp==0.11.0
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
rich==13.9.4
six==1.17.0
smolagents==1.6.0
soupsieve==2.6
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
webencodings==0.5.1
yfinance==0.2.52
